+++ run_tests.py
@@ -8,7 +8,7 @@ from tests.review_test import CarSelectionTest, LoginTest, LogoutTest, AverageRa
 
 if __name__ == '__main__':
     suite = unittest.TestSuite((
-        # unittest.makeSuite(AddReviewErrorsTest)
+        unittest.makeSuite(AddReviewErrorsTest),
         # unittest.makeSuite(LoginTest),
         # unittest.makeSuite(LogoutTest),
         # unittest.makeSuite(AverageRatingTest),
