+++ tests/buy_tests.py
@@ -1,0 +1,5 @@
+from tests.base import BaseTestCase
+
+
+class AddReviewErrorsTest(BaseTestCase):
+    pass
